+++ traffic_monitor/tools/testto/Dockerfile_run.sh
@@ -15,4 +15,19 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-go build -ldflags "-X main.GitRevision=`git rev-parse HEAD` -X main.BuildTimestamp=`date +'%Y-%M-%dT%H:%M:%S'`"
+
+start() {
+  ARGS=
+  if [[ -n "${PORT}" ]]; then
+    ARGS="$ARGS -port ${PORT}"
+  fi
+  testto ${ARGS}
+}
+
+init() {
+  echo "INITIALIZED=1" >>/etc/environment
+}
+
+source /etc/environment
+if [ -z "$INITIALIZED" ]; then init; fi
+start